*: clarify license terms
I've relicensed everything (hopefully for the last time) such that all
code is under the GNU GPLv3 or later, and any other files are under the
MIT license (because I don't believe that configuration should be
copyrightable). Also I updated the copyright years to be correct (and
added headers to files where they were missing).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>