yomichan: custom anki template
I don't like how the definitions are styled by default (with "|"
separators) so use "," instead. I would also like to be able to get rid
of the "JMDict (English)" tag as well but doing so would also remove the
tag for 国語辞典 -- where I do actually care about the name of the
dictionary.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>