mpv: clean up script fetching and maintenance setup
It's not the best idea to to use submodules for mpv scripts, it's better
to maintain them using an after-install hook because that way we ensure
that we keep the git repo functional.

Also we don't need sub-pause-skip anymore since mpvacious supports this
feature now (not that I used it a lot anyway).

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>