keyboards: use i3 config for configuration
It turns out there's a bunch of bugs that make configuring keyboards
with xorg.conf.d in anything other than the most trivial setups
effectively impossible. xprofile also appears to not work properly, so
the only remaining option is to use the i3 startup configuration.

In addition, there appear to be issues with xinput and trying to have
both multi- and single- layout configured keyboards at the same time.
So, I have to (unfortunately) just live with having to switch layouts
manually with RAlt+RShift.

This is all very unfortunate.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>