mutt: sandbox mailcap-triggered programs
This will block most trivial attempts to exfiltrate data (through
tracking links and images via the network), but this is not an ideal
solution as our entire /home is still accessible to the program.

Ideally we would limit the scope (maybe with AppArmor, or even just
masking /home with a tmpfs inside the sandbox userns) but that can be
done later.

Signed-off-by: Aleksa Sarai <cyphar@cyphar.com>